pyproject.toml
[project]
name = "data-lakehouse"
version = "0.1.0"
description = "Modern data lakehouse with medallion architecture using NYC Taxi data, DuckDB, and dbt"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Abdullah Al Harbi" }]
keywords = ["data-lakehouse", "dbt", "duckdb", "medallion-architecture", "data-engineering"]

dependencies = [
    "dbt-duckdb>=1.9.0",
    "httpx>=0.27.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
